+++ demo/feature_test.py
@@ -12,7 +12,7 @@ rb.init()
 main = rb.Scene()
 rb.Game.scenes.add(main, "main")
 
-main.add(rb.Sprite({"pos": rb.Vector(100, 100)}).add(rb.Circle({"color": rb.Color.green})))
+main.add(rb.GameObject({"pos": rb.Vector(100, 100)}).add(rb.Circle({"color": rb.Color.green})))
 
 vec = rb.Vector(10, 10).dir_to(rb.Vector(11, 20))
 vec.round(3)