+++ docs/source/conf.py
@@ -153,17 +153,6 @@ napoleon_use_ivar = False
 napoleon_use_param = True
 napoleon_use_rtype = True
 napoleon_preprocess_types = True
-napoleon_type_aliases = {
-    "SceneManager": ":meth:`SceneManager <rubato.scenes.scene_manager.SceneManager>`",
-    "Radio": ":meth:`Radio <rubato.radio.Radio>`",
-    "STATE": ":meth:`STATE <rubato.utils.STATE.STATE>`",
-    "Game": ":meth:`Game <rubato.game.Game>`",
-    "Vector": ":meth:`Vector <rubato.utils.vector.Vector>`",
-    "Camera": ":meth:`Camera <rubato.scenes.camera.Camera>`",
-    "Polygon": ":meth:`Polygon <rubato.utils.sat.Polygon>`",
-    "COL_TYPE": ":meth:`COL_TYPE <rubato.utils.COL_TYPE.COL_TYPE>`",
-    "Image": ":meth:`Image <rubato.sprite.image.Image>`",
-}
 napoleon_attr_annotations = True
 
 # Version Control