+++ rubato/__init__.py
@@ -1,6 +1,6 @@
 """
 Rubato is a modern 2D game engine for python. Accurate fixed-step physics
-simulations, robust scene and sprite management, event listener system and more
+simulations, robust scene and game object management, event listener system and more
 all come prepackaged.
 
 Fundamentally, Rubato is built developer-focused. From intricate rigidbody
@@ -22,7 +22,7 @@ from rubato.game import Game
 from rubato.utils import Math, Display, Vector, Time, Color, Defaults
 from rubato.utils.error import *
 from rubato.radio import Radio
-from rubato.classes import SceneManager, Scene, Camera, Sprite, Image
+from rubato.classes import SceneManager, Scene, Camera, GameObject, Image, Spritesheet
 from rubato.classes import RigidBody, Animation, Component, Polygon, Rectangle, Circle, ColInfo, SAT, Hitbox, Group
 import rubato.input as Input
 from rubato.sound import Sound
@@ -51,7 +51,7 @@ __all__ = [
     "Scene",
     "Camera",
     "Radio",
-    "Sprite",
+    "GameObject",
     "Image",
     "RigidBody",
     "Input",
@@ -62,6 +62,7 @@ __all__ = [
     "Hitbox",
     "Group",
     "Game",
+    "Spritesheet",
 ]
 
 
@@ -78,23 +79,27 @@ def init(options: dict = {}):
 
     params = Defaults.game_defaults | options
 
-    Game.border_color = Color(
-        *params["border_color"]) if not isinstance(params["border_color"], Color) else params["border_color"]
+    Game.border_color = Color(*params["border_color"]
+                             ) if not isinstance(params["border_color"], Color) else params["border_color"]
 
-    Game.background_color = Color(*params["background_color"]) if not isinstance(params["background_color"],
-                                                                                 Color) else params["background_color"]
+    Game.background_color = Color(*params["background_color"]
+                                 ) if not isinstance(params["background_color"], Color) else params["background_color"]
 
     Time.target_fps = params["target_fps"]
     Time.physics_fps = params["physics_fps"]
 
-    flags = (sdl2.SDL_WINDOW_RESIZABLE | sdl2.SDL_WINDOW_ALLOW_HIGHDPI | sdl2.SDL_WINDOW_SHOWN |
-             sdl2.SDL_WINDOW_MOUSE_FOCUS | sdl2.SDL_WINDOW_INPUT_FOCUS)
+    flags = (
+        sdl2.SDL_WINDOW_RESIZABLE | sdl2.SDL_WINDOW_ALLOW_HIGHDPI | sdl2.SDL_WINDOW_SHOWN |
+        sdl2.SDL_WINDOW_MOUSE_FOCUS | sdl2.SDL_WINDOW_INPUT_FOCUS
+    )
 
     Display.window = sdl2.ext.Window(params["name"], params["window_size"].to_tuple(), flags=flags)
 
-    Display.renderer = sdl2.ext.Renderer(Display.window,
-                                         flags=(sdl2.SDL_RENDERER_ACCELERATED | sdl2.SDL_RENDERER_PRESENTVSYNC),
-                                         logical_size=params["res"].to_tuple())
+    Display.renderer = sdl2.ext.Renderer(
+        Display.window,
+        flags=(sdl2.SDL_RENDERER_ACCELERATED | sdl2.SDL_RENDERER_PRESENTVSYNC),
+        logical_size=params["res"].to_tuple()
+    )
 
     if params["icon"] != "":
         Display.set_window_icon(params["icon"])
@@ -114,4 +119,5 @@ def begin():
         Game.constant_loop()
     else:
         raise RuntimeError(
-            "You have not initialized rubato. Make sure to run rubato.init() right after importing the library")
+            "You have not initialized rubato. Make sure to run rubato.init() right after importing the library"
+        )