+++ rubato/classes/__init__.py
@@ -3,9 +3,10 @@ The sprite folder holds all of the classes dealing with sprites.
 Everything here works with rubato.sprite.Sprite.
 """
 from rubato.classes.camera import Camera
-from rubato.classes.sprite import Sprite
 from rubato.classes.component import Component
 from rubato.classes.components import *
+from rubato.classes.game_object import GameObject
+from rubato.classes.spritesheet import Spritesheet
 from rubato.classes.group import Group
 from rubato.classes.scene import Scene
 from rubato.classes.scene_manager import SceneManager