+++ rubato/classes/components/__init__.py
@@ -1,5 +1,5 @@
 """
-The types folder holds specific sprite subclasses.
+Holds all the various components for game objects.
 """
 from rubato.classes.components.hitbox import Hitbox, Polygon, Rectangle, Circle, SAT, ColInfo
 from rubato.classes.components.image import Image