+++ rubato/utils/error.py
@@ -20,7 +20,7 @@ class SideError(Exception):
 
 class DuplicateComponentError(Exception):
     """
-    An error that is raised when you try to add a component to a sprite
+    An error that is raised when you try to add a component to a game object
     that already has a component of the same type
     """
     pass
@@ -28,7 +28,7 @@ class DuplicateComponentError(Exception):
 
 class ComponentNotAllowed(Exception):
     """
-    An error that is raised when you try to add a component on a sprite that
-    is not allowed by another component on that sprite.
+    An error that is raised when you try to add a component on a game object that
+    is not allowed by another component on that game object.
     """
     pass